refactor functions into sepr. files
