+++ .vscode/settings.json
@@ -1,5 +1,5 @@
 {
   "editor.defaultFormatter": "esbenp.prettier-vscode",
   "editor.formatOnSave": true,
-  "editor.formatOnPaste": false
- }
+  "editor.formatOnPaste": true
+}
