+++ wrangler.toml-EXAMPLE
@@ -1,3 +1,3 @@
-name = "bot-tg"
+name = "<WORKER_PREFIX_BEFORE_YOUR_DOMAIN>"
 main = "src/index.ts"
 compatibility_date = "2023-03-30"
